Added better comments
